Fix for dicom with float images (apply_voi_lut)
Hello,

We've just received new dicom where `'RescaleSlope' in ds and 'RescaleIntercept' in ds`

Just for your understanding, all I was doing is:
```
# reading dcm
dcm = pydicom.read_file("/home/alexander/work/reader_test/7FD1838B")

# extracting image (pixels data)
img = apply_voi_lut(apply_modality_lut(dcm.pixel_array,dcm),dcm)

plt.imshow(img)
```
So I dont do any conversion myself. Just pydicom things.

If you pay attention to changed line, clipped_iv should never be float. So I was thinking about adding error handling that will say "array should be int", but instead I just decided to init zeros array with proper dtype, now it works fine

+++ pydicom/pixel_data_handlers/util.py
@@ -334,7 +334,7 @@ def apply_voi_lut(arr, ds, index=0):
         lut_data = np.asarray(lut_data, dtype=dtype)
 
         # IVs < `first_map` get set to first LUT entry (i.e. index 0)
-        clipped_iv = np.zeros(arr.shape, dtype=arr.dtype)
+        clipped_iv = np.zeros(arr.shape, dtype=dtype)
         # IVs >= `first_map` are mapped by the VOI LUT
         # `first_map` may be negative, positive or 0
         mapped_pixels = arr >= first_map
